Fix some know issues.
1. If a disk's superblock is not clear, an incorrect md device will still be assembled.
I can use dd command to clear MBR, but we also have to notify user about the risks of deleting a volume.
Force to clear when the result of Examine_ToResult is not EXAMINE_NO_MD_SUPERBLOCK
2. MD device is still being formatted but user requeste to delete it at the same time.
Check whether MD device is in formatting state or not before starting deleting procedures.