Aggregate symbolic link, disk type and MD device information into MiscDiskInfo structure.
Because the information has a very high probability to change according to
the system's status and user's operations. To manage them conveniently,
I decide to move them out and create a new structure for them.

m_miscInfo in RAIDDiskInfo won't always need to update. I will set it
when some processes want to get the disk information.