Unmount a volume won't clear volume number.
I want to keep the volume number to show on Web UI.
So, I added a new method, GetFormerVolumeNum, which can try to get
former allocated volume number for the volume. However, we will free
the number when the volume is unmouted, so the number may occupy by
other volumes. Therefore, the method will try to get another new number
unless there is no free number can be assigned.