Move workable MD device checking forward.
RAID01, RAID1, RAID5 and RAID6 can afford some disk lost.
If the check does not move forward, at initial stage, some MD devices
won't be mounted due to the code only allows the MD with the active
and spare disks' number equals to the disk number when RAID is created.