Add event callback register/deregister methods and define callback events.
1. As title. Those methods are provided to the module who want
RAIDManager to callback when some RAID events happen.

2. Remove the critical protection and create thread call in constructor. I
consider that critical protection is unnecessary because the module is
in initial stage and it should be no one to send request at the time.
And I remove create thread call to provide the flexibility to the using
module to decide the timing it wants to start the monitoring thread.