Handle the event set outside the thread.
Add a new event, CB_EVENT_DELRAID_DONE, which is used to notify other
modules that RAID deletion is done. This event is raised outside the
thread. So, we add the handle logic in the thread. Only the monitoring
thread can issue the callback to the module who registered the callback
function.